requirements.txt
apturl==0.5.2
asgiref==3.5.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
dell-recovery==0.0.0
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
progressbar==2.5
protobuf==3.12.4
proton-client==0.6.1
protonvpn-cli==3.9.1
protonvpn-gui==1.3.0
protonvpn-nm-lib==3.4.1
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python3-openid==3.2.0
pythondialog==3.5.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sos==4.3
SQLAlchemy==1.4.31
sqlparse==0.4.2
system-service==0.3
systemd-python==234
ttystatus==0.38
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlrd==1.2.0
zipp==1.0.0
